+++ scripts/run_db_seeder_generate_schema.sh
@@ -22,7 +22,7 @@ fi
 echo "================================================================================"
 echo "Start $0"
 echo "--------------------------------------------------------------------------------"
-echo "DB Seeder - Generation of database schema."
+echo "DBSeeder - Generation of database schema."
 echo "--------------------------------------------------------------------------------"
 echo "FILE_CONFIGURATION_NAME   : ${DB_SEEDER_FILE_CONFIGURATION_NAME}"
 echo "FILE_JSON_NAME            : ${DB_SEEDER_FILE_JSON_NAME}"