+++ Release-Notes.md
@@ -131,13 +131,15 @@ n/a
 
 #### Modified Features
 
+- CockroachDB: DBMS v21.1.2
+
 - Exasol: DBMS 7.0.10
 
 - xxx: DBMS 999 / JDBC 999
 
 #### Deleted Features
 
-n/a
+- Docker Compose functionality removed
 
 ----------
 
@@ -172,8 +174,6 @@ Release Date: 04.06.2021
 - trino: DBMS 358 / JDBC 358
 
 - YugabyteDB: DBMS 2.7.1.1-b1
-
-- xxx: DBMS 999 / JDBC 999
 
 ----------